# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test app | Secunda test",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/buildings": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Получить список зданий",
                "operationId": "704d1f7897ed7a3b8d29dacea66823ad",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 1000,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Building"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Создать новое здание",
                "operationId": "c8f8399919e01b1e4e2e2f6a9f1f5107",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BuildingCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Здание создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Building"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/buildings/{building}": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Получить здание по ID",
                "operationId": "e96dddf69feb8b2f20edde85d7850827",
                "parameters": [
                    {
                        "name": "building",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Building"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено"
                    }
                }
            },
            "put": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Обновить здание",
                "operationId": "e493ff77b33d03a5021738d3a2ef197d",
                "parameters": [
                    {
                        "name": "building",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BuildingUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Здание обновлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Building"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено"
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Удалить здание",
                "operationId": "564a6a0f37a2da5ef23bb80f6d79c6f7",
                "parameters": [
                    {
                        "name": "building",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Здание удалено"
                    },
                    "404": {
                        "description": "Здание не найдено"
                    }
                }
            }
        },
        "/api/buildings/{building}/organizations": {
            "get": {
                "tags": [
                    "Buildings",
                    "Organizations"
                ],
                "summary": "Получить список организаций в здании",
                "description": "Возвращает пагинированный список всех организаций, принадлежащих конкретному зданию",
                "operationId": "922fc30c8107019ce3f8acb6f90d988c",
                "parameters": [
                    {
                        "name": "building",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Название организации"
                                                    },
                                                    "phones": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "8012421412"
                                                        }
                                                    },
                                                    "building_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-09-25T15:44:53.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-09-25T15:44:54.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Building not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Building": {
                "title": "Building",
                "description": "Модель здания/строения",
                "required": [
                    "description",
                    "address",
                    "longitude",
                    "latitude"
                ],
                "properties": {
                    "id": {
                        "description": "ID здания",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "description": {
                        "description": "Описание здания",
                        "type": "string",
                        "example": "Офисное здание бизнес-центра"
                    },
                    "address": {
                        "description": "Адрес здания",
                        "type": "string",
                        "example": "ул. Примерная, д. 123"
                    },
                    "longitude": {
                        "description": "Долгота",
                        "type": "number",
                        "format": "float",
                        "example": 55.751244
                    },
                    "latitude": {
                        "description": "Широта",
                        "type": "number",
                        "format": "float",
                        "example": 37.618423
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "BuildingCreate": {
                "title": "Building Create",
                "description": "Данные для создания здания",
                "required": [
                    "description",
                    "address",
                    "longitude",
                    "latitude"
                ],
                "properties": {
                    "description": {
                        "description": "Описание здания",
                        "type": "string",
                        "example": "Офисное здание бизнес-центра"
                    },
                    "address": {
                        "description": "Адрес здания",
                        "type": "string",
                        "example": "ул. Примерная, д. 123"
                    },
                    "longitude": {
                        "description": "Долгота",
                        "type": "number",
                        "format": "float",
                        "example": 55.751244
                    },
                    "latitude": {
                        "description": "Широта",
                        "type": "number",
                        "format": "float",
                        "example": 37.618423
                    }
                },
                "type": "object"
            },
            "BuildingUpdate": {
                "title": "Building Update",
                "description": "Данные для обновления здания",
                "properties": {
                    "description": {
                        "description": "Описание здания",
                        "type": "string",
                        "example": "Обновленное описание здания"
                    },
                    "address": {
                        "description": "Адрес здания",
                        "type": "string",
                        "example": "ул. Обновленная, д. 456"
                    },
                    "longitude": {
                        "description": "Долгота",
                        "type": "number",
                        "format": "float",
                        "example": 55.755826
                    },
                    "latitude": {
                        "description": "Широта",
                        "type": "number",
                        "format": "float",
                        "example": 37.6173
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost/api/buildings"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "to": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost/api/buildings?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost/api/buildings?page=10"
                    },
                    "prev": {
                        "type": "string",
                        "example": "http://localhost/api/buildings?page=1"
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost/api/buildings?page=2"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "description": "Постоянный API ключ для доступа к API",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}